requirements.txt
discord
requests
mysql-connector-python